# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.4"
  title: "Byzanti.ne API Gateway - Powered by Liberty Block"
# during dev, should point to your local machine
host: local.byzanti.ne:8901
...
basePath: /
securityDefinitions:
  # X-API-Key: abcdef12345
  APIKeyQueryParam:
    type: apiKey
    in: query
    name: api_key
  APIKeyHeaderParam:
    type: apiKey
    in: header
    name: api_key
  # APIKeyCookieParam:
  #   type: apiKey
  #   in: cookie
  #   name: api_key
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json

# Global security (applies to all operations):
security:
  - APIKeyQueryParam: []
  - APIKeyHeaderParam: []
  # - APIKeyCookieParam: []
paths:
...
  /createAccount:
    x-swagger-router-controller: createAccount
    post:
      tags:
        - Base API {account,tokens,transaction,actions}
      description: Create an EOS Account
      operationId: createAccount
      parameters:
        - name: body
          in: body
          description: Body should encapsulate {creator=EOS account name creating the account}, {name=name of the account being created},{owner=Owner Public key},{active=Active Public key},{sig=account creator signature}
          required: true
          schema:
            properties:
              creator:
                type: string
              name:
                type: string
              owner:
                type: string
              active:
                type: string
              sig:
                type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/transferResponse"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: Error
          schema:
            $ref: "#/definitions/TransferErrorResponse"
...
  /voteProducer:
    x-swagger-router-controller: voteProducer
    post:
      tags:
        - Block Producer API {Producer List, Voting, Delegate, Undelegate, Buy/sell RAM & get RAM price} *mainnet
      description: Lets you vote for a specific producer account
      operationId: voteProducer
      parameters:
        - name: body
          in: body
          description: Body should encapsulate {voter=Account to transfer from}, {producer=comma seperated values of producers to vote for},{sig=voter signature}
          required: true
          schema:
            properties:
              voter:
                type: string
              producer:
                type: string
              sig:
                type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/transferResponse"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: Error
          schema:
            $ref: "#/definitions/TransferErrorResponse"
  /delegate:
    x-swagger-router-controller: delegate
    post:
      tags:
        - Block Producer API {Producer List, Voting, Delegate, Undelegate, Buy/sell RAM & get RAM price} *mainnet
      description: Delegate specific NET and CPU bandwidth for a user
      operationId: delegate
      parameters:
        - name: body
          in: body
          description: Body should encapsulate {from=account delegating bandwidth}, {receiver=account to delegate bandwidth to},{net=amount of EOS to undelegate for network bandwidth},{cpu=amount of EOS to undelegate for CPU bandwidth},{sig=voter signature}
          required: true
          schema:
            properties:
              from:
                type: string
              receiver:
                type: string
              net:
                type: string
              cpu:
                type: string
              sig:
                type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/transferResponse"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: Error
          schema:
            $ref: "#/definitions/TransferErrorResponse"
  /undelegate:
    x-swagger-router-controller: undelegate
    post:
      tags:
        - Block Producer API {Producer List, Voting, Delegate, Undelegate, Buy/sell RAM & get RAM price} *mainnet
      description: Undelegate specific NET and CPU bandwidth
      operationId: undelegate
      parameters:
        - name: body
          in: body
          description: Body should encapsulate {from=account undelegating bandwidth}, {receiver=account to undelegate bandwidth from},{net=amount of EOS to undelegate for network bandwidth},{cpu=amount of EOS to undelegate for CPU bandwidth},{sig=voter signature}
          required: true
          schema:
            properties:
              from:
                type: string
              receiver:
                type: string
              net:
                type: string
              cpu:
                type: string
              sig:
                type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/transferResponse"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: Error
          schema:
            $ref: "#/definitions/TransferErrorResponse"
...
  /buyRamBytes:
    x-swagger-router-controller: buyRamBytes
    post:
      tags:
        - Block Producer API {Producer List, Voting, Delegate, Undelegate, Buy/sell RAM & get RAM price} *mainnet
      description: Buy Ram in specified bytes
      operationId: buyRamBytes
      parameters:
        - name: body
          in: body
          description: Body should encapsulate {payer=account paying for RAM}, {receiver=account receiving RAM},{bytes=RAM in bytes},{sig=payer signature}
          required: true
          schema:
            properties:
              payer:
                type: string
              receiver:
                type: string
              bytes:
                type: integer
              sig:
                type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/transferResponse"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: Error
          schema:
            $ref: "#/definitions/TransferErrorResponse"
...
definitions:
...
  ErrorResponse:
    properties:
      error:
        type: string
  TransferErrorResponse:
    properties:
      error:
        type: string
...
  transferResponse:
    type: object
    items:
      type: object
      example: { object: value }